added functionality to update sample data with form
